refactor(proxy): split proxy_failover.go by concern
proxy_failover.go held the candidate attempt loop alongside two
self-contained pieces. Same package, pure moves with their doc comments:

- upstream_error_forward.go: forwardableErrorStatus,
  forwardableErrorBodyCap, forwardUpstreamError, the credential masker
  (keyShapedToken, credentialMinLen, credentialMasker, exactMaskWriter,
  maskKeyShapedTokens) and carriesErrorObject
- param_retry.go: responsesLearnBodyCap, paramRetryResult,
  learnedScopeFor, candidateModelID, retryLearnable400,
  learnRejectedParams, paramRetryRounds, issueParamRetry,
  readLearnable400, retryWithStrippedParams, renameKeys

proxy_failover.go keeps attemptCandidate, failAllExhausted,
classifyProbeFailure, dispatchStreaming, beginAttempt,
touchProviderLastUsed, buildCandidateRequest, resolveCancelOrigin,
doUpstream and recordBreakerOutcome (1380 -> 725 lines) and leaves the
size allowlist. Tests are unchanged.
