+++ internal/proxy/proxy_failover.go
@@ -3,14 +3,11 @@ package proxy
 import (
 	"bytes"
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
 	"net/http"
 	"net/http/httptrace"
-	"regexp"
-	"slices"
 	"strings"
 	"sync/atomic"
 	"time"
@@ -41,41 +38,6 @@ import (
 // error message and well below the multi-megabyte bodies the image endpoints can
 // answer with.
 const failoverErrorClassifyCap = 16 << 10
-
-// responsesLearnBodyCap bounds a 400 that has to be parsed rather than scanned.
-//
-// openairesponses.RequiresResponsesAPI json.Unmarshals the whole error document,
-// so the classifier's window is the wrong size for it: a body cut off mid-JSON
-// does not parse, and the /v1/responses requirement would silently stop being
-// learned. A megabyte is far past any real 400 and still bounded.
-const responsesLearnBodyCap = 1 << 20
-
-// paramRetryResult is the outcome of the 400 param-stripping auto-retry
-// (retryWithStrippedParams). It tells the failover loop how to proceed:
-//   - resp: the response to continue handling with — the last retry's response
-//     once any retry was issued, otherwise the original 400. Every 400 that
-//     leaves here carries a restored body, so normal non-200 handling can read
-//     it whether it is the original or the one a retry earned.
-//   - retryCancel: the retry context's cancel func, non-nil only when a retry
-//     response is live and its body has NOT yet been consumed. The caller must
-//     call it after consuming the body.
-//   - streamCancelOrigin: "retry_timeout" once a retry was issued, otherwise
-//     the caller's original value, unchanged.
-//   - retried: true once a retry request was issued and answered, whatever it
-//     answered with — the caller must fold the retries' dial time into the
-//     running totals.
-//   - lastReqErr: set only when cont is true; the structured cause the caller
-//     records via st.setReqErr before failing over.
-//   - cont: true => the caller should `continue` to the next candidate (a retry
-//     request could not be created or failed).
-type paramRetryResult struct {
-	resp               *http.Response
-	retryCancel        context.CancelFunc
-	streamCancelOrigin string
-	retried            bool
-	lastReqErr         reqError
-	cont               bool
-}
 
 // failAllExhausted handles phase E: every candidate failed (or the overall
 // deadline was hit). It logs the exhaustion, records a 502 failure row, and
@@ -719,325 +681,6 @@ func (h *Handler) doUpstream(ctx context.Context, req *http.Request, st *request
 	return resp, true
 }
 
-// forwardableErrorStatus reports whether a status is in the payload class a
-// client may see the provider's body for: a 4xx that judged the caller's own
-// request. Deliberately static where shouldFailover is dynamic: its 429
-// verdict follows the failover_on_rate_limit setting, but a quota body is the
-// operator's account state whichever way that toggle points, and what can
-// reach a client must not be a side effect of a routing knob. The denied 4xx
-// are the auth, billing, quota and routing classes whose bodies can carry
-// operator account detail; 1xx/3xx are not payload errors and 5xx bodies are
-// the provider talking about itself, so none of those forward either.
-func forwardableErrorStatus(status int) bool {
-	if status < 400 || status >= 500 {
-		return false
-	}
-	switch status {
-	case http.StatusUnauthorized, http.StatusPaymentRequired, http.StatusForbidden,
-		http.StatusNotFound, http.StatusProxyAuthRequired, http.StatusTooManyRequests, 499:
-		return false
-	}
-	return true
-}
-
-// forwardableErrorBodyCap bounds a payload-class error body that may be
-// forwarded to the client. Reading it whole keeps forwarded JSON intact, but
-// "whole" needs a ceiling: the multimodal endpoints have no upstream pre-cap,
-// so a broken or hostile custom endpoint could answer an error status with
-// anything. A megabyte is far past any real error document; a body over it is
-// answered with the synthesised envelope instead.
-const forwardableErrorBodyCap = 1 << 20
-
-// forwardUpstreamError handles a non-200 upstream response that is NOT being
-// failed over (phase G): log + meter the failure via failRequest, then answer the
-// client. isFailoverEligible carries the caller's shouldFailover verdict; what
-// the client may see is decided by it together with the static
-// forwardableErrorStatus class. A payload-class refusal (a plain 400 and its
-// kin) judged this caller's own request, so the upstream's error object is
-// forwarded with the provider's credential masked (credentialMasker: exact
-// key, then key-shaped tokens). Everything else - auth, billing,
-// rate limit, not-found, server faults, whether classed by eligibility or by
-// status - gets a synthesised envelope with the classified reason, because
-// those bodies can quote the operator's provider credentials or account
-// details; the body stays in the request log either way, with only this
-// gateway's own key redacted. Drains/closes
-// resp.Body exactly once and always returns outcomeFatal.
-func (h *Handler) forwardUpstreamError(w http.ResponseWriter, st *requestState, candidate modelCandidate, resp *http.Response, attempt int, isFailoverEligible bool, responseHeaderMs float64) candidateOutcome {
-	logData := st.logData
-	masker := logData.masker
-	mayForwardError := !isFailoverEligible && forwardableErrorStatus(resp.StatusCode)
-	// How much of the body is worth holding depends on what happens to it
-	// below. A 2xx is forwarded whole - truncating a success would corrupt it.
-	// A forwardable error is read under its own cap, and one that overflows it
-	// is demoted to the envelope rather than forwarded truncated, so a client
-	// never receives invalid JSON where the provider sent something complete.
-	// A discarded body is read under the same cap as the two drain sites,
-	// since all that is left to take from it is a classification and the first
-	// 10 000 bytes of request log.
-	var body []byte
-	oversized := false
-	switch {
-	case resp.StatusCode/100 == 2:
-		body, _ = io.ReadAll(resp.Body)
-	case mayForwardError:
-		body, _ = io.ReadAll(io.LimitReader(resp.Body, forwardableErrorBodyCap+1))
-		if len(body) > forwardableErrorBodyCap {
-			oversized = true
-			body = body[:forwardableErrorBodyCap]
-		}
-		_, _ = io.Copy(io.Discard, resp.Body)
-	default:
-		body, _ = io.ReadAll(io.LimitReader(resp.Body, failoverErrorClassifyCap))
-		_, _ = io.Copy(io.Discard, resp.Body)
-	}
-	_ = resp.Body.Close()
-	errMsg := util.SanitizeLogBody(string(body), 10000)
-	// Classify for the request log and metrics only — routing is unaffected,
-	// the caller already decided it from the status code.
-	kind, reason := classifyUpstreamError(resp.StatusCode, errMsg, candidate.model.ModelID)
-	if kind == KindProviderModelGone {
-		// Same as the drain path above: the candidate carries what the
-		// pre-retirement probe needs, and logData.endpointType is the family
-		// that decides whether this model can be adjudicated at all.
-		h.noteModelGone(candidate, logData.endpointType)
-	}
-	debuglog.Warn("proxy: upstream non-200", "status", resp.StatusCode, "error_kind", kind, "model", logData.modelID, "provider", logData.providerName, "provider_id", candidate.provider.ID)
-	debuglog.Debug("proxy: upstream error response", "status", resp.StatusCode, "model", logData.modelID, "provider", logData.providerName, "provider_id", candidate.provider.ID, "body_length", len(body), "attempt", attempt+1)
-	logData.responseHeaderMs = responseHeaderMs
-	h.failRequest(logData, resp.StatusCode, kind, string(masker.mask([]byte(errMsg))), attempt, st.startTime, st.parseMs, st.timings, st.cacheHits, st.proxyOverhead)
-
-	// A 2xx that reached this function (any success status other than a bare
-	// 200) is not an error at all and is forwarded whatever its body is,
-	// including a non-JSON or empty one. That is what lets a 204 answer with
-	// no body: an envelope written under a No Content status would be a body
-	// this gateway invented for a request the provider considered successful.
-	// Checked before eligibility so a success can never be rewritten.
-	if resp.StatusCode/100 == 2 {
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(resp.StatusCode)
-		//nolint:gosec // G705 false positive: provider JSON body, not HTML; Content-Type is application/json
-		_, _ = w.Write(masker.maskExact(body))
-		return outcomeFatal
-	}
-
-	if !mayForwardError {
-		// Auth, billing, rate-limit, not-found and server-fault classes -
-		// whether ruled out by the caller's eligibility verdict or by the
-		// static status class. Their bodies are the ones that can quote the
-		// operator's provider credentials ("Incorrect API key provided:
-		// sk-...") or account details a virtual-key holder must not see, so the
-		// body (own key redacted) stays in the DB request log via failRequest and the caller gets
-		// the classified reason: enough to tell "this model is gone" from "top
-		// up your account" from "try again shortly".
-		writeOpenAIError(w, upstreamClientMessage(candidate.provider.Name, resp.StatusCode, reason), resp.StatusCode)
-		return outcomeFatal
-	}
-
-	// Payload-class refusal: the provider judged this caller's own request, so
-	// the caller is entitled to the detail, whether or not this was the last
-	// candidate.
-	switch {
-	case carriesErrorObject(body) && !oversized:
-		// Forward the upstream response so clients can react to semantic errors
-		// (e.g. context_length_exceeded). The upstream's own error object
-		// carries detail this gateway cannot reconstruct — code, type, param,
-		// provider-specific fields — so it is passed through byte for byte
-		// apart from masking key-shaped tokens, since even a payload error is
-		// provider-authored free text.
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(resp.StatusCode)
-		//nolint:gosec // G705 false positive: provider JSON error body, not HTML; Content-Type is application/json
-		_, _ = w.Write(masker.mask(body))
-	case json.Valid(body):
-		// A non-2xx whose JSON body carries no error object. OpenCode Zen and
-		// OpenCode Go answer some failed requests with a complete
-		// chat.completion envelope under an HTTP 400, which is valid JSON with
-		// nothing for a client to read `.error.message` off. There is no
-		// upstream error detail to preserve here, so the classified reason is
-		// synthesised into an envelope instead; the body itself stays in the
-		// request log via failRequest above.
-		writeOpenAIError(w, upstreamClientMessage(candidate.provider.Name, resp.StatusCode, reason), resp.StatusCode)
-	default:
-		// Body is not JSON (e.g. HTML from a CDN). Wrap in an
-		// OpenAI-compatible envelope so JSON-parsing clients don't crash.
-		//
-		// The sanitized body rides inside the message here, where the case above
-		// hands back only the classified reason. The asymmetry is deliberate: the
-		// full body reaches the request log either way via failRequest, and how
-		// much of a provider's response this gateway echoes to callers is one
-		// decision for all three cases rather than something to widen here.
-		writeOpenAIError(w, string(masker.mask([]byte(errMsg))), resp.StatusCode)
-	}
-	return outcomeFatal
-}
-
-// keyShapedToken matches credential-looking substrings a provider may quote
-// inside an error body: prefixed secret keys (sk- also covers sk-ant-, sk-or-
-// and sk-proj-; hf_, fw_, r8_, gsk_, xai- cover HuggingFace, Fireworks,
-// Replicate, Groq and xAI), Google API keys (AIza...), AWS access key ids
-// (AKIA...), bare JWTs (the MiniMax API key format), and bearer tokens. The
-// minimum tail lengths keep prose like "sk-abc" out of scope; matches without
-// a digit are prose too and are dropped by maskKeyShapedTokens. A prefix list
-// necessarily trails the provider roster - it is the third layer, behind the
-// status-class gate above and credentialMasker's exact match of this
-// gateway's own key.
-var keyShapedToken = regexp.MustCompile(`\b(?:sk|gsk|xai|hf|fw|r8)[-_][A-Za-z0-9_-]{16,}|\bAIza[0-9A-Za-z_-]{30,}|\bAKIA[0-9A-Z]{16}\b|\beyJ[A-Za-z0-9_-]{10,}\.[A-Za-z0-9_-]{10,}\.[A-Za-z0-9_-]{5,}|(?i:\bbearer\s+)[A-Za-z0-9._~+/=-]{16,}`)
-
-// credentialMinLen is the shortest provider key the exact-value mask will
-// redact. Keyless local providers carry an empty key and a handful of test or
-// placeholder setups use tiny ones; rewriting every occurrence of a few
-// characters would shred the body without protecting anything.
-const credentialMinLen = 8
-
-// credentialMasker scrubs a provider's credential out of client-bound text.
-// The exact decrypted key is the primary control: it is an exact byte match,
-// so it covers every key shape, including custom or self-hosted gateways
-// whose format the prefix regex can never anticipate, but not a JSON-escaped
-// rendering of the key (an encoder turning "&" into "\u0026" or "/" into
-// "\/" defeats it; real keys rarely carry such bytes). maskKeyShapedTokens
-// runs after it as the third layer, behind the status-class gate, for
-// credentials a relay quotes that are not ours. Build one per candidate with
-// newCredentialMasker and pass it to every client-facing emit of provider
-// error text: the buffered paths in forwardUpstreamError and the in-stream
-// error frames on the translated (handleDataChunk), native Anthropic
-// (emitRawData) and pass-through (sseErrorMaskWriter) streaming paths. The
-// zero value masks by shape only. The exact layer alone (maskExact) also runs
-// on every other provider body a client receives; the request log's error
-// message gets both layers, since it is error text readable by non-admin
-// users with the logs grant. Content bodies never meet the regex.
-type credentialMasker struct {
-	secret []byte
-}
-
-func newCredentialMasker(apiKey string) credentialMasker {
-	if len(apiKey) < credentialMinLen {
-		return credentialMasker{}
-	}
-	return credentialMasker{secret: []byte(apiKey)}
-}
-
-// mask returns body with every occurrence of the exact credential, then any
-// key-shaped token, replaced by "[redacted]". The exact pass runs first so the
-// regex cannot split the key and leave a recognisable remainder. For error
-// frames and bodies only: the regex layer can match prose.
-func (m credentialMasker) mask(body []byte) []byte {
-	return maskKeyShapedTokens(m.maskExact(body))
-}
-
-// maskExact replaces only the exact credential. It cannot false-positive, so
-// it is safe on every provider body bound for a client (content chunks,
-// success bodies) and on the request log, where it closes the read a
-// VK-owning dashboard user has on their own failed requests.
-func (m credentialMasker) maskExact(body []byte) []byte {
-	if len(m.secret) > 0 && bytes.Contains(body, m.secret) {
-		return bytes.ReplaceAll(body, m.secret, []byte("[redacted]"))
-	}
-	return body
-}
-
-// maskKeyShapedTokens scrubs credential-looking substrings from upstream error
-// text before it reaches a client or the request log. Auth-class errors never
-// forward at all, and credentialMasker has already removed this gateway's own
-// key by exact value; this is the third layer, for a provider quoting some
-// other credential inside an otherwise forwardable payload error. A match with no
-// digit in it is an identifier or prose ("sk_business_unit_identifier",
-// "Bearer authentication-required") rather than a credential, and stays -
-// real keys carry digits. The replacement carries no JSON metacharacters, so
-// a valid body stays valid.
-// exactMaskWriter applies credentialMasker.maskExact to a byte stream whose
-// writes may split the key: it holds back the last len(key)-1 bytes of each
-// write until the next one arrives, so a key straddling two writes is still
-// seen whole. Flush releases the held tail at end of stream. Used on the two
-// raw forwarding paths (an oversized pass-through JSON remainder and an
-// oversized SSE event) where per-chunk masking has boundaries.
-type exactMaskWriter struct {
-	w    io.Writer
-	cred credentialMasker
-	tail []byte
-}
-
-func newExactMaskWriter(w io.Writer, cred credentialMasker) *exactMaskWriter {
-	return &exactMaskWriter{w: w, cred: cred}
-}
-
-func (e *exactMaskWriter) Write(p []byte) (int, error) {
-	if len(e.cred.secret) == 0 {
-		return e.w.Write(p)
-	}
-	buf := e.cred.maskExact(slices.Concat(e.tail, p))
-	keep := min(len(e.cred.secret)-1, len(buf))
-	out := buf[:len(buf)-keep]
-	e.tail = append([]byte(nil), buf[len(buf)-keep:]...)
-	if len(out) > 0 {
-		if _, err := e.w.Write(out); err != nil {
-			return 0, err
-		}
-	}
-	return len(p), nil
-}
-
-// Flush writes the held tail. Call once the stream (or the raw stretch of
-// it) has ended.
-func (e *exactMaskWriter) Flush() error {
-	if len(e.tail) == 0 {
-		return nil
-	}
-	out := e.tail
-	e.tail = nil
-	_, err := e.w.Write(out)
-	return err
-}
-
-func maskKeyShapedTokens(body []byte) []byte {
-	return keyShapedToken.ReplaceAllFunc(body, func(m []byte) []byte {
-		if !bytes.ContainsAny(m, "0123456789") {
-			return m
-		}
-		return []byte("[redacted]")
-	})
-}
-
-// carriesErrorObject reports whether an upstream body is a JSON object with an
-// "error" member that actually carries something, which is what decides between
-// forwarding that body verbatim and synthesising an envelope over it.
-//
-// The test is emptiness, not shape. What a provider puts inside its error is not
-// this gateway's to judge, so any populated value counts: an object with fields,
-// Ollama's bare string ("model not found"), a list, even a number. What does not
-// count is a member that leaves a client with nothing to read - `null`, `{}`,
-// `""`, `[]` - because a body carrying one of those is, from the caller's side,
-// the same body with no error member at all, which is exactly the case this
-// function exists to catch. A body that is not a JSON object (an array, a bare
-// string, HTML) can carry no member and reports false.
-func carriesErrorObject(body []byte) bool {
-	var envelope map[string]json.RawMessage
-	if err := json.Unmarshal(body, &envelope); err != nil {
-		return false
-	}
-	raw, present := envelope["error"]
-	if !present {
-		return false
-	}
-	var content any
-	if err := json.Unmarshal(raw, &content); err != nil {
-		return false
-	}
-	switch v := content.(type) {
-	case nil:
-		return false
-	case string:
-		return strings.TrimSpace(v) != ""
-	case map[string]any:
-		return len(v) > 0
-	case []any:
-		return len(v) > 0
-	default:
-		// A number or a bool: peculiar, but the provider put a value there and
-		// a client can render it.
-		return true
-	}
-}
-
 // recordBreakerOutcome records the circuit-breaker result for a completed
 // upstream attempt (phase D8). It is a no-op when the breaker is disabled.
 //
@@ -1079,302 +722,4 @@ func (h *Handler) recordBreakerOutcome(st *requestState, candidate modelCandidat
 	if !st.isStreaming || statusCode != http.StatusOK {
 		h.circuitBreaker.RecordSuccess(candidate.provider.ID, candidate.provider.Name)
 	}
-}
-
-// learnedScopeFor scopes the learned reject/rename caches to one provider. The
-// provider id, not its type: many distinct custom endpoints share the "openai"
-// type, and a type-scoped entry would let one of them disable a param for all
-// the others serving the same model id.
-func learnedScopeFor(candidate modelCandidate) string {
-	if candidate.provider == nil {
-		return ""
-	}
-	return candidate.provider.ID.String()
-}
-
-// candidateModelID reads a candidate's upstream model id for diagnostics.
-// recordBreakerOutcome is reached on paths that only carry the provider, so the
-// model must never be assumed present.
-func candidateModelID(candidate modelCandidate) string {
-	if candidate.model == nil {
-		return ""
-	}
-	return candidate.model.ModelID
-}
-
-// retryLearnable400 picks the self-heal that fits what this attempt actually
-// sent, and reports handled=false when none does. Each family can only read the
-// dialect it speaks: a 400 names the fields of the request that earned it, so a
-// reading taken from the wrong dialect is not merely useless but harmful — a
-// param mislearned from a Messages 400 would poison the compat path for that
-// model on every later request.
-//
-// Chat-completions attempts get two, in this order: retryWithResponses runs
-// BEFORE the param retry because a Responses-required 400 names
-// reasoning_effort, which the param parser would otherwise learn as a strip
-// (useless on reason-by-default models). The param retry itself works
-// universally — any LLM API naming "temperature" or "top_p" in a 400 can only
-// mean the sampling parameter — but it rebuilds the OpenAI-shaped st.bodyBytes
-// and re-POSTs it to the same targetURL, which for a native /v1/messages,
-// /v1/responses or generateContent route would be a malformed request.
-//
-// Anthropic egress attempts get the one 400 that route can fix by asking
-// differently: a model refusing the extended-thinking shape it was asked in
-// (anthropic_thinking_retry.go). Every other Messages 400 is left alone.
-func (h *Handler) retryLearnable400(
-	r *http.Request,
-	st *requestState,
-	candidate modelCandidate,
-	providerType, targetURL string,
-	resp *http.Response,
-	attempt int,
-	dialMs *float64,
-	failoverCancel context.CancelFunc,
-	streamCancelOrigin string,
-) (paramRetryResult, bool) {
-	switch {
-	case st.anthropicEgressAttempt:
-		return h.retryLearnableMessages400(r, st, candidate, providerType, resp, attempt, dialMs, failoverCancel, streamCancelOrigin)
-	case st.sentChatCompletionsBody():
-		res, handled := h.retryWithResponses(r, st, candidate, providerType, resp, attempt, dialMs, failoverCancel, streamCancelOrigin)
-		if !handled {
-			res = h.retryWithStrippedParams(r, st, candidate, providerType, targetURL, resp, attempt, dialMs, failoverCancel, streamCancelOrigin)
-		}
-		return res, true
-	}
-	return paramRetryResult{resp: resp, streamCancelOrigin: streamCancelOrigin}, false
-}
-
-// learnRejectedParams caches the params and renames a 400 body names, so later
-// requests to the same provider+model are built without them. It is the
-// learning half of retryWithStrippedParams, split out for callers that cannot
-// retry in place (a hedged probe, which must not spend a second round-trip
-// inside one race slot).
-func (h *Handler) learnRejectedParams(candidate modelCandidate, body []byte) {
-	rejected := paramrewrite.ParseProviderParamError(body)
-	renames := paramrewrite.ParseProviderParamRename(body)
-	if rejected == nil && renames == nil {
-		return
-	}
-	cacheKey := paramrewrite.LearnedCacheKey(learnedScopeFor(candidate), candidate.model.ModelID)
-	if rejected != nil {
-		paramrewrite.MergeLearnedParamCache(&h.deprecationCache, cacheKey, rejected)
-	}
-	if renames != nil {
-		paramrewrite.MergeLearnedParamCache(&h.paramRenameCache, cacheKey, renames)
-	}
-	debuglog.Info("proxy: learned rejected params from upstream 400", "provider", candidate.provider.Name, "model", candidate.model.ModelID, "rejected", fmt.Sprintf("%v", rejected), "renames", fmt.Sprintf("%v", renames))
-}
-
-// paramRetryRounds caps how many times one candidate is re-issued with newly
-// learned params stripped.
-//
-// Upstreams name a single offending param per 400, so a request carrying two of
-// them needs a second round to get through. Past that the provider is objecting
-// to something this self-heal cannot fix, and the 400 belongs to the client.
-const paramRetryRounds = 2
-
-// issueParamRetry rebuilds the upstream body with strip applied on top of the
-// learned caches and re-POSTs it to targetURL.
-//
-// The returned cancel func belongs to the retry's context: non-nil exactly when
-// a response is returned, whose body the caller must consume before calling it.
-// On failure the context is cancelled here (there is no body to read) and the
-// structured cause is returned for the failover loop to record.
-func (h *Handler) issueParamRetry(
-	r *http.Request,
-	st *requestState,
-	candidate modelCandidate,
-	providerType, targetURL string,
-	strip map[string]bool,
-	attempt int,
-	dialMs *float64,
-) (*http.Response, context.CancelFunc, *reqError) {
-	// Rebuild the request body using the shared rewrite path. This ensures
-	// stream_options injection, provider injection, universal/learned param
-	// stripping, and learned renaming are all applied on retry, preventing drift
-	// from the initial attempt path. The renames just cached are picked up from
-	// paramRenameCache; the rejected params learned so far are also passed as
-	// extraStrip so the retry drops them even before the cache writes are observed.
-	rebuilt := paramrewrite.BuildUpstreamBody(st.bodyBytes, providerType, candidate.model.ModelID, st.reqModel, st.isStreaming, &h.deprecationCache, &h.paramRenameCache, strip, learnedScopeFor(candidate))
-	retryCtx, rc := context.WithTimeout(r.Context(), st.failoverTimeout)
-	retryCtx = context.WithValue(retryCtx, ctxkeys.CancelOriginKey, "retry_timeout")
-	retryCtx = context.WithValue(retryCtx, ctxkeys.DialMsKey, dialMs)
-	retryReq, retryErr := newRequestWithContext(retryCtx, "POST", targetURL, bytes.NewReader(rebuilt))
-	if retryErr != nil {
-		rc()
-		return nil, nil, &reqError{Kind: KindInternal, Attempt: attempt, Provider: candidate.provider.Name, Underlying: errString(retryErr)}
-	}
-	util.SetProviderAuthHeaders(retryReq, providerType, candidate.apiKey)
-	retryReq.Header.Set("Content-Type", "application/json")
-	var retryCheckRedirect func(req *http.Request, via []*http.Request) error
-	if h.safeDialer != nil {
-		retryCheckRedirect = h.safeDialer.CheckRedirect
-	}
-	retryClient := &http.Client{Transport: h.upstreamTransport, CheckRedirect: retryCheckRedirect}
-	//nolint:bodyclose // retryResp.Body is returned to the caller, which consumes and closes it
-	retryResp, retryErr := retryClient.Do(retryReq)
-	if retryErr != nil {
-		rc() // no body to consume on retry error
-		debuglog.Warn("proxy: auto-retry request failed", "attempt", attempt+1, "provider", candidate.provider.Name, "provider_id", candidate.provider.ID, "error", retryErr)
-		if errors.Is(retryErr, context.Canceled) || errors.Is(retryErr, context.DeadlineExceeded) {
-			// Branch like the main failover loop: Canceled = client
-			// disconnect, DeadlineExceeded = retry timeout.
-			origin := "retry_timeout"
-			if errors.Is(retryErr, context.Canceled) {
-				origin = "client_disconnect"
-			}
-			return nil, nil, &reqError{Kind: cancelOriginToKind(origin), Attempt: attempt, Provider: candidate.provider.Name}
-		}
-		return nil, nil, &reqError{Kind: KindProviderError, Attempt: attempt, Provider: candidate.provider.Name, Underlying: errString(retryErr)}
-	}
-	return retryResp, rc, nil
-}
-
-// readLearnable400 consumes a 400 body for the param self-heal: it reads what
-// the learner can parse, drains the rest so the connection stays reusable,
-// closes the original body and restores a readable one in its place — the
-// response is handed on to whoever renders the client's error, so it must never
-// leave here empty.
-//
-// The bound is the parse-sized responsesLearnBodyCap rather than the
-// scan-sized failoverErrorClassifyCap: learning json.Unmarshals the whole
-// document, so a body cut mid-JSON parses to nothing and teaches nothing. The
-// cap sits far past any real provider error, and everything above it is
-// discarded rather than held.
-func readLearnable400(resp *http.Response) ([]byte, error) {
-	body, err := io.ReadAll(io.LimitReader(resp.Body, responsesLearnBodyCap))
-	_, _ = io.Copy(io.Discard, resp.Body)
-	_ = resp.Body.Close()
-	resp.Body = io.NopCloser(bytes.NewReader(body))
-	return body, err
-}
-
-// retryWithStrippedParams handles a 400 from an upstream: it reads and restores
-// the error body, cancels the original (now-finished) request context, and — if
-// the body is a recognizable param-rejection — learns the rejected params and
-// renames into the deprecation/rename caches, rebuilds the request with them
-// applied, and re-issues it. A retry that is itself rejected with a 400 is read
-// and learned from in exactly the same way, and re-issued once more when it
-// names a param the retry did not already carry. See paramRetryResult for how
-// the loop interprets the return value.
-//
-// failoverCancel is the original request's cancel func; it is invoked here once
-// the 400 body has been consumed. dialMs is the per-request dial-timing pointer
-// threaded into every retry context so SafeDialer records the retries' DNS time.
-func (h *Handler) retryWithStrippedParams(
-	r *http.Request,
-	st *requestState,
-	candidate modelCandidate,
-	providerType, targetURL string,
-	resp *http.Response,
-	attempt int,
-	dialMs *float64,
-	failoverCancel context.CancelFunc,
-	streamCancelOrigin string,
-) paramRetryResult {
-	// Restored before anything else, so downstream error handling can read the
-	// body on every path that gives up without retrying.
-	body, readErr := readLearnable400(resp)
-	failoverCancel() // 400 body consumed, context no longer needed
-	debuglog.Debug("proxy: received 400 from upstream, checking for param rejection", "provider", candidate.provider.Name, "provider_id", candidate.provider.ID, "model", candidate.model.ModelID, "body_length", len(body))
-
-	res := paramRetryResult{resp: resp, streamCancelOrigin: streamCancelOrigin}
-	if readErr != nil {
-		return res
-	}
-
-	// Everything applied to the outgoing body so far, accumulated across rounds:
-	// strip holds the rejected params (dropped), renamed the moves (value kept
-	// under the new name). They are also the termination guard — a round is only
-	// worth issuing when a 400 names something outside these sets.
-	strip := map[string]bool{}
-	renamed := map[string]string{}
-
-	// A 400 can ask us to drop a param (rejected → strip) and/or move a param to
-	// a new name (rename → preserve value). Both feed the same self-heal: learn,
-	// cache for future preemptive application, and retry. learnFrom does the
-	// learning half for one 400 body and reports whether that body named anything
-	// the request does not already carry — i.e. whether re-issuing could help.
-	learnFrom := func(errBody []byte) bool {
-		rejected := paramrewrite.ParseProviderParamError(errBody)
-		renames := paramrewrite.ParseProviderParamRename(errBody)
-		if rejected == nil && renames == nil {
-			return false
-		}
-		// Cache the learned rejections and renames for future preemptive
-		// application. Each cache is merged with any existing entries via
-		// CompareAndSwap to avoid data races from concurrent goroutines mutating
-		// the same map.
-		h.learnRejectedParams(candidate, errBody)
-		progressed := false
-		for p := range rejected {
-			if !strip[p] {
-				progressed = true
-			}
-			strip[p] = true
-		}
-		for oldName, newName := range renames {
-			if _, seen := renamed[oldName]; !seen {
-				progressed = true
-			}
-			renamed[oldName] = newName
-		}
-		return progressed
-	}
-
-	// Two guards, both required, so the loop always terminates: at most
-	// paramRetryRounds rounds, and a round only when the current 400 names
-	// something the request does not already carry (learnFrom's report), which
-	// makes strip∪renamed grow strictly on every iteration.
-	//
-	// learnFrom runs before the round check on purpose — it is the left operand,
-	// so it is evaluated even on the iteration that ends the loop. That is what
-	// makes the last 400 learned rather than discarded: the round it would have
-	// paid for is refused, but the next request still starts with what it named.
-	//
-	// The first 400 that names nothing recognisable falls straight out here and
-	// is handed back untouched.
-	for round := 0; learnFrom(body) && round < paramRetryRounds; round++ {
-		res.streamCancelOrigin = "retry_timeout"
-		retryResp, rc, retryErr := h.issueParamRetry(r, st, candidate, providerType, targetURL, strip, attempt, dialMs)
-		if retryErr != nil {
-			res.lastReqErr = *retryErr
-			res.cont = true
-			return res
-		}
-		if retryResp.StatusCode != http.StatusBadRequest {
-			// rc must NOT be called here — retryResp.Body is read by the caller.
-			// It is returned for deferred cleanup after body consumption.
-			res.resp = retryResp
-			res.retryCancel = rc
-			res.retried = true
-			debuglog.Info("proxy: auto-retry succeeded", "model", candidate.model.ModelID, "rounds", round+1, "rejected_params", mapKeys(strip), "renamed_params", renameKeys(renamed))
-			return res
-		}
-		// The retry was rejected too. Read its body so the params it names are
-		// learned as well, then hand that response on with the body restored: it
-		// is the client's 400 if no further round is issued.
-		body, readErr = readLearnable400(retryResp)
-		rc() // retry body consumed and buffered, context no longer needed
-		res.resp = retryResp
-		res.retryCancel = nil
-		res.retried = true
-		if readErr != nil {
-			return res
-		}
-	}
-	return res
-}
-
-// renameKeys returns the old param names from a rename map, for log fields.
-func renameKeys(m map[string]string) []string {
-	if len(m) == 0 {
-		return nil
-	}
-	keys := make([]string, 0, len(m))
-	for k := range m {
-		keys = append(keys, k)
-	}
-	return keys
 }
