+++ scripts/ci/size-allowlist.txt
@@ -28,7 +28,6 @@ internal/proxy/model_gone_test.go	2739
 internal/proxy/models_test.go	2307
 internal/proxy/multimodal.go	1071
 internal/proxy/proxy.go	976
-internal/proxy/proxy_failover.go	1380
 internal/proxy/upstream_classify.go	804
 web/src/api/types.ts	1058
 web/src/components/Layout.tsx	1065
